`update 80/200 v13`

+++ commands/Fun/reverse.js
@@ -10,31 +10,17 @@ module.exports = {
     async execute(client, message, cmd, args, Discord) {
         const text = args.join(" ")
         if (!text) {
-            const nopr = new Discord.MessageEmbed()
-                .setTimestamp()
-                .setColor(`${color}`)
-                .setAuthor(`${message.author.username}`, message.author.displayAvatarURL({ dynamic: true }))
-                .setDescription(`**\`(prefix)reverse <text>\`**`)
-            return message.lineReplyNoMention(nopr)
+            const nopr = new Discord.MessageEmbed().setTimestamp().setColor(`${color}`).setAuthor(`${message.author.username}`, message.author.displayAvatarURL({ dynamic: true })).setDescription(`**\`(prefix)reverse <text>\`**`)
+            return message.lineReplyNoMention({ embed: nopr })
         }
         if (message.content.match("erog") || message.content.match("boob") || message.content.match("stit") || message.content.match("kcuf") || message.content.match("tihs") || message.content.match("aggin") || message.content.match("eoh") || message.content.match("hctib") || message.content.match("kcid") || message.content.match("yag") || message.content.match("naibsel") || message.content.match("bojwolb") || message.content.match("nrop") || message.content.match("tnuc") || message.content.match("@")) {
-            message.delete();
-            const embed1 = new Discord.MessageEmbed()
-                .setTimestamp()
-                .setColor(`${color}`)
-                .setAuthor(`${message.author.username}`, message.author.displayAvatarURL({ dynamic: true }))
-                .setDescription(`**Im Not Allowed To Say Bad Words Or Ping! Lol Cry 😂**`)
-            message.lineReplyNoMention(embed1)
+            message.delete(); const embed1 = new Discord.MessageEmbed().setTimestamp().setColor(`${color}`).setAuthor(`${message.author.username}`, message.author.displayAvatarURL({ dynamic: true })).setDescription(`**Im Not Allowed To Say Bad Words Or Ping! Lol Cry 😂**`)
+            message.lineReplyNoMention({ embed: embed1 })
         }
-
         let Rarray = text.split("")
         let reverseArray = Rarray.reverse()
         let result = reverseArray.join("")
-        const embed = new Discord.MessageEmbed()
-            .setTimestamp()
-            .setColor(`${color}`)
-            .setTitle('Reverse Conversion')
-            .setDescription(`**${result}**`)
-        message.lineReplyNoMention(embed)
+        const embed = new Discord.MessageEmbed().setTimestamp().setColor(`${color}`).setTitle('Reverse Conversion').setDescription(`**${result}**`)
+        message.lineReplyNoMention({ embed: embed })
     }
 }
